+++ rhonn/__init__.py
@@ -1,0 +1,6 @@
+import rhonn
+
+import importlib.metadata
+
+
+__version__ = importlib.metadata.version("rhonn")
